+++ lib/sunzi.rb
@@ -1,28 +1,21 @@
-require 'thor'
-require 'rainbow'
+# Stdlib
+require 'erb'
+require 'pathname'
 require 'yaml'
 
-# Starting 2.0.0, Rainbow no longer patches string with the color method by default.
-require 'rainbow/version'
-require 'rainbow/ext/string' unless Rainbow::VERSION < '2.0.0'
+# Gems
+require 'hashugar'
+require 'rainbow/ext/string'
+require 'thor'
 
-module Sunzi
-  autoload :Cli,        'sunzi/cli'
-  autoload :Cloud,      'sunzi/cloud'
-  autoload :Dependency, 'sunzi/dependency'
-  autoload :DNS,        'sunzi/dns'
-  autoload :Logger,     'sunzi/logger'
-  autoload :Utility,    'sunzi/utility'
+# Sunzi
+require 'sunzi/utility'
+require 'sunzi/worker'
+require 'sunzi/command'
+require 'sunzi/cli'
+require 'sunzi/dependency'
+require 'sunzi/logger'
 
-  class Cloud
-    autoload :Base,         'sunzi/cloud/base'
-    autoload :Linode,       'sunzi/cloud/linode'
-    autoload :DigitalOcean, 'sunzi/cloud/digital_ocean'
-  end
-
-  class DNS
-    autoload :Base,     'sunzi/dns/base'
-    autoload :Linode,   'sunzi/dns/linode'
-    autoload :Route53,  'sunzi/dns/route53'
-  end
-end
+# Plug-ins
+require 'sunzi/plugin'
+Sunzi::Plugin.load