+++ lib/sunzi/dependency.rb
@@ -1,39 +1,37 @@
 module Sunzi
-  class Dependency
-    def self.all
-      {
-        'linode' =>   { :require => 'linode',   :version => '>= 0.7.9' },
-        'highline' => { :require => 'highline', :version => '>= 1.6.11'},
-        'route53' =>  { :require => 'route53',  :version => '>= 0.2.1' },
-        'digital_ocean' =>  { :require => 'digital_ocean',  :version => '>= 1.0.0' },
-      }
+  class Dependency < Struct.new(:name, :version)
+
+    @list = {}
+
+    def initialize(*args)
+      super
+      self.class.list[name] = self
     end
 
-    def self.load(name)
-      begin
-        gem(name, all[name][:version])
-        require(all[name][:require])
-      rescue LoadError
-        if $!.to_s =~ /Gemfile/
-          Logger.error <<-EOS
-Dependency missing: #{name}
-Add this line to your application's Gemfile.
+    class << self
+      attr_accessor :list
 
-    gem '#{name}', '#{all[name][:version]}'
-
-Please try again after running "bundle install".
-          EOS
-        else
-          Logger.error <<-EOS
-Dependency missing: #{name}
-To install the gem, issue the following command:
-
-    gem install #{name} -v '#{all[name][:version]}'
-
-Please try again after installing the missing dependency.
-          EOS
+      def load(key)
+        unless dependency = @list[key]
+          fail "#{key} is not initialized. Run Sunzi::Dependency.new('#{key}', '~> ...')"
         end
-        abort
+
+        name, version = dependency.name, dependency.version
+
+        begin
+          gem(name, version)
+          require(name)
+        rescue LoadError
+          base = Pathname.new(__FILE__).dirname.parent.join('templates/dependency')
+          which = if $!.to_s =~ /Gemfile/
+            'gemfile'
+          else
+            'install'
+          end
+          text = ERB.new(base.join("#{which}.erb").read, nil, '-').result(binding)
+          Logger.error text
+          abort
+        end
       end
     end
   end