+++ lib/sunzi/dns/base.rb
@@ -1,7 +1,0 @@
-module Sunzi
-  class DNS
-    class Base
-      include Sunzi::Utility
-    end
-  end
-end