+++ test/test_helper.rb
@@ -1,0 +1,6 @@
+require 'sunzi'
+require 'minitest/autorun'
+
+require 'pathname'
+
+GemRoot = Pathname.new(__FILE__).dirname.parent